Added Model Structure Property
Changes:
- Moved metadata/modeling from `StructureProperty` to new `ModelStructureProperty`
- Improved parsing of structure vals to better use the models they are bound to
- Added support for saving nested properties when saving the structure property
- Added customizable "sql_type" to pick among the TEXT data types
